# ecommerce/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/catagory": {
            "get": {
                "description": "get all catagory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show all catagory product ecommerce",
                "operationId": "get-allcatagory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsCat"
                        }
                    },
                    "400": {
                        "description": "Id is not exsis\"}",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsFalse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsFalse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsFalse"
                        }
                    }
                }
            }
        },
        "/mitra/product": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get all product mitra in mitra",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show all My Product",
                "operationId": "get-all myproduct mitra",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Respons"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsFalse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsFalse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponsFalse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Catagory": {
            "type": "object",
            "properties": {
                "catagory_name": {
                    "type": "string",
                    "example": "Electronik"
                },
                "user_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "model.Respons": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string",
                    "example": "Success"
                },
                "Success": {
                    "type": "boolean",
                    "example": true
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "model.ResponsCat": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string",
                    "example": "Success"
                },
                "Success": {
                    "type": "boolean",
                    "example": true
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Catagory"
                    }
                }
            }
        },
        "model.ResponsFalse": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string",
                    "example": "message"
                },
                "Success": {
                    "type": "boolean",
                    "example": false
                },
                "data": {
                    "type": "object"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    },
    "x-extension-openapi": {
        "example": "value on a json format"
    }
}